[fix]デプロイ持エラーのためpura.rbを編集

+++ db/migrate/20250911044329_add_rank_id_and_next_kente_rank_id_to_students.rb
@@ -1,7 +1,7 @@
 class AddRankIdAndNextKenteRankIdToStudents < ActiveRecord::Migration[6.1]
   def change
-    add_column :students, :rank_id, :integer
-    add_column :students, :next_kente_rank_id, :integer
+    add_column :students, :rank_id, :bigint
+    add_column :students, :next_kente_rank_id, :bigint
 
     # 外部キー制約を追加 (KenteRankモデルがある場合)
     add_foreign_key :students, :kente_ranks, column: :rank_id
